Added publish/subscribe design pattern. Cleaned up code in other patterns.
