+++ DesignPatternsCSharpNet6/Command/Pattern/Account.cs
@@ -1,0 +1,13 @@
+﻿namespace DesignPatternsCSharpNet6.Command.Pattern;
+
+public class Account
+{
+    public string OwnerName { get; set; }
+    public decimal Balance { get; set; }
+
+    public Account(string ownerName, decimal balance)
+    {
+        OwnerName = ownerName;
+        Balance = balance;
+    }
+}
