+++ DesignPatternsCSharpNet6/Command/Pattern/Enums.cs
@@ -1,0 +1,9 @@
+﻿namespace DesignPatternsCSharpNet6.Command.Pattern;
+
+public enum ExecutionStatus
+{
+    Unprocessed,
+    InsufficientFunds,
+    ExecuteFailed,
+    ExecuteSucceeded
+}
