+++ DesignPatternsCSharpNet6/Command/Pattern/ITransaction.cs
@@ -1,0 +1,7 @@
+﻿namespace DesignPatternsCSharpNet6.Command.Pattern;
+
+public interface ITransaction
+{
+    ExecutionStatus Status { get; set; }
+    void Execute();
+}
