+++ DesignPatternsCSharpNet6/Command/Pattern/Withdraw.cs
@@ -1,0 +1,31 @@
+﻿namespace DesignPatternsCSharpNet6.Command.Pattern;
+
+public class Withdraw : ITransaction
+{
+    private readonly Account _account;
+    private readonly decimal _amount;
+
+    public ExecutionStatus Status { get; set; }
+
+    public Withdraw(Account account, decimal amount)
+    {
+        _account = account;
+        _amount = amount;
+
+        Status = ExecutionStatus.Unprocessed;
+    }
+
+    public void Execute()
+    {
+        if(_account.Balance >= _amount)
+        {
+            _account.Balance -= _amount;
+
+            Status = ExecutionStatus.ExecuteSucceeded;
+        }
+        else
+        {
+            Status = ExecutionStatus.InsufficientFunds;
+        }
+    }
+}
