--- DesignPatternsCSharpNet6/Command/PatternVersion/Account.cs
@@ -1,13 +1,0 @@
-﻿namespace DesignPatternsCSharpNet6.Command.PatternVersion;
-
-public class Account
-{
-    public string OwnerName { get; set; }
-    public decimal Balance { get; set; }
-
-    public Account(string ownerName, decimal balance)
-    {
-        OwnerName = ownerName;
-        Balance = balance;
-    }
-}
