--- DesignPatternsCSharpNet6/Command/PatternVersion/Deposit.cs
@@ -1,24 +1,0 @@
-﻿namespace DesignPatternsCSharpNet6.Command.PatternVersion;
-
-public class Deposit : ITransaction
-{
-    private readonly Account _account;
-    private readonly decimal _amount;
-
-    public ExecutionStatus Status { get; set; }
-
-    public Deposit(Account account, decimal amount)
-    {
-        _account = account;
-        _amount = amount;
-
-        Status = ExecutionStatus.Unprocessed;
-    }
-
-    public void Execute()
-    {
-        _account.Balance += _amount;
-
-        Status = ExecutionStatus.ExecuteSucceeded;
-    }
-}
