--- DesignPatternsCSharpNet6/Command/PatternVersion/Enums.cs
@@ -1,9 +1,0 @@
-﻿namespace DesignPatternsCSharpNet6.Command.PatternVersion;
-
-public enum ExecutionStatus
-{
-    Unprocessed,
-    InsufficientFunds,
-    ExecuteFailed,
-    ExecuteSucceeded
-}
