--- DesignPatternsCSharpNet6/Command/PatternVersion/ITransaction.cs
@@ -1,7 +1,0 @@
-﻿namespace DesignPatternsCSharpNet6.Command.PatternVersion;
-
-public interface ITransaction
-{
-    ExecutionStatus Status { get; set; }
-    void Execute();
-}
