+++ DesignPatternsCSharpNet6/DesignPatternsCSharpNet6.csproj
@@ -8,7 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Builder\" />
     <Folder Include="DataMapper\" />
   </ItemGroup>
