+++ DesignPatternsCSharpNet6/PublishSubscribe/NonPattern/Models/Player.cs
@@ -1,0 +1,13 @@
+﻿namespace DesignPatternsCSharpNet6.PublishSubscribe.NonPattern.Models;
+
+public class Player
+{
+    public string Name { get; }
+    public int HitPoints { get; set; }
+
+    public Player(string name, int hitPoints)
+    {
+        Name = name;
+        HitPoints = hitPoints;
+    }
+}
