+++ DesignPatternsCSharpNet6/PublishSubscribe/Pattern/Models/Location.cs
@@ -1,0 +1,13 @@
+﻿namespace DesignPatternsCSharpNet6.PublishSubscribe.Pattern.Models;
+
+public class Location
+{
+    public string Name { get; }
+    public bool IsPoisonous { get; }
+
+    public Location(string name, bool isPoisonous)
+    {
+        Name = name;
+        IsPoisonous = isPoisonous;
+    }
+}
