+++ DesignPatternsCSharpNet6/Strategy/Pattern/AverageByMean.cs
@@ -1,0 +1,9 @@
+﻿namespace DesignPatternsCSharpNet6.Strategy.Pattern;
+
+public class AverageByMean : IAveragingStrategy
+{
+    public double CalculateAverage(List<double> values)
+    {
+        return values.Sum() / values.Count;
+    }
+}
