+++ DesignPatternsCSharpNet6/Strategy/Pattern/IAveragingStrategy.cs
@@ -1,0 +1,6 @@
+﻿namespace DesignPatternsCSharpNet6.Strategy.Pattern;
+
+public interface IAveragingStrategy
+{
+    double CalculateAverage(List<double> values);
+}
