--- DesignPatternsCSharpNet6/Strategy/PatternVersion/AverageByMean.cs
@@ -1,9 +1,0 @@
-﻿namespace DesignPatternsCSharpNet6.Strategy.PatternVersion;
-
-public class AverageByMean : IAveragingStrategy
-{
-    public double CalculateAverage(List<double> values)
-    {
-        return values.Sum() / values.Count;
-    }
-}
