--- DesignPatternsCSharpNet6/Strategy/PatternVersion/AverageByMedian.cs
@@ -1,17 +1,0 @@
-﻿namespace DesignPatternsCSharpNet6.Strategy.PatternVersion;
-
-public class AverageByMedian : IAveragingStrategy
-{
-    public double CalculateAverage(List<double> values)
-    {
-        var sortedValues = values.OrderBy(x => x).ToList();
-
-        if(sortedValues.Count % 2 == 1)
-        {
-            return sortedValues[(sortedValues.Count - 1) / 2];
-        }
-
-        return (sortedValues[(sortedValues.Count / 2) - 1] + 
-                sortedValues[sortedValues.Count / 2]) / 2;
-    }
-}
