--- DesignPatternsCSharpNet6/Strategy/PatternVersion/IAveragingStrategy.cs
@@ -1,6 +1,0 @@
-﻿namespace DesignPatternsCSharpNet6.Strategy.PatternVersion;
-
-public interface IAveragingStrategy
-{
-    double CalculateAverage(List<double> values);
-}
