+++ Test.DesignPatternsCSharpNet6/BuilderPattern/NonPatternVersion/TestNonPatternReport.cs
@@ -1,4 +1,4 @@
-﻿using DesignPatternsCSharpNet6.Builder.NonPatternVersion;
+﻿using DesignPatternsCSharpNet6.Builder.NonPattern;
 
 namespace Test.DesignPatternsCSharpNet6.BuilderPattern.NonPatternVersion;
